+++ cfgov/data_research/forms.py
@@ -5,3 +5,12 @@ class ConferenceRegistrationForm(forms.ModelForm):
     class Meta:
         model = ConferenceRegistration
         fields = '__all__'
+
+    def __init__(self, *args, **kwargs):
+        super(ConferenceRegistrationForm, self).__init__(*args, **kwargs)
+        self.fields['name'].required = True
+        self.fields['sessions'].required = True
+        self.fields['codes'].required = True
+        self.fields['sessions'].error_messages.update({
+            'required': 'You must select at least one session to attend.'
+        })