+++ cfgov/data_research/migrations/0001_initial.py
@@ -17,9 +17,10 @@ class Migration(migrations.Migration):
                 ('name', models.CharField(max_length=250, blank=True)),
                 ('organization', models.CharField(max_length=250, blank=True)),
                 ('email', models.EmailField(max_length=250, blank=True)),
-                ('sessions', models.TextField(blank=True)),
+                ('sessions', models.TextField()),
                 ('foodinfo', models.CharField(max_length=250, blank=True)),
                 ('accommodations', models.CharField(max_length=250, blank=True)),
+                ('codes', models.TextField()),
             ],
         ),
     ]